0731 14:48 login, join 수정중

+++ src/main/java/net/dima/dima5_project/entity/PortPredictEntity.java
@@ -25,7 +25,7 @@ public class PortPredictEntity {
     
     @Id
     @Column(name="predict_seq")
-    private int predictSeq;
+    private Long predictSeq;
 
     // @Column(name="port_id")
     // private String portId;
